extract-get-draft-kv-pool-move: Move get_draft_kv_pool to mem_cache.kv_cache_builder
Mechanical cut + paste for the ``extract-get-draft-kv-pool`` mech move.

Cut ``Scheduler.get_draft_kv_pool`` (a @staticmethod after the prep
commit) and paste it as a module-level free function in
``python/sglang/srt/mem_cache/kv_cache_builder.py`` (new
package). Drop ``@staticmethod`` decorator; body bytes unchanged.

Caller sites updated: ``Scheduler.get_draft_kv_pool(...)`` →
``kv_cache_builder.get_draft_kv_pool(...)`` (pure prefix replacement). Add import
``from sglang.srt.mem_cache import kv_cache_builder``.

Verify via ``git --color-moved-ws=allow-indentation-change``: the entire
function body should be marked as moved.

Refactor chain ID: extract-get-draft-kv-pool-move
